JBIDE-16538 - initial commit of bpel publisher refactor
JBIDE-16538 missed some files

JBIDE-16538
