+++ plugins/org.jboss.tools.bpel.runtimes/META-INF/MANIFEST.MF
@@ -9,6 +9,7 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
  org.eclipse.core.resources,
  org.eclipse.wst.server.core,
+ org.eclipse.wst.server.ui,
  org.eclipse.wst.web.ui,
  org.eclipse.wst.common.project.facet.ui,
  org.eclipse.wst.common.modulecore,
@@ -16,20 +17,18 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.wst.common.project.facet.core,
  org.eclipse.wst.common.frameworks.ui,
  org.eclipse.jst.server.generic.core,
+ org.eclipse.jst.common.frameworks,
  org.eclipse.jem.util,
  org.eclipse.debug.ui,
  org.eclipse.bpel.model,
  org.eclipse.emf.ecore,
  org.eclipse.wst.wsdl,
- javax.wsdl,
- org.eclipse.jst.common.frameworks,
- org.jboss.ide.eclipse.as.wtp.core,
  org.eclipse.wst.web,
- org.jboss.ide.eclipse.as.core,
- org.jboss.ide.eclipse.archives.webtools,
- org.jboss.tools.jmx.core,
- org.eclipse.wst.server.ui,
- org.eclipse.ui.navigator
+ org.eclipse.ui.navigator,
+ javax.wsdl,
+ org.jboss.tools.foundation.core,
+ org.jboss.ide.eclipse.as.wtp.core;bundle-version="3.0.0",
+ org.jboss.ide.eclipse.as.core;bundle-version="3.0.0"
 Eclipse-LazyStart: true
 Export-Package: org.jboss.tools.bpel.runtimes,
  org.jboss.tools.bpel.runtimes.facets,
