+++ plugins/org.jboss.tools.bpel.runtimes/plugin.properties
@@ -29,3 +29,7 @@ NEW_BPEL_PROJECT_DESCRIPTION=Create a new BPEL project (for SOA-P v5.1 and earli
 OLD_BPEL_CORE_FACET_LABEL=JBoss BPEL 2.0 Facet (SOA-P v5.1 and earlier)
 OLD_BPEL_CORE_FACET_DESCRIPTION=Enables modeling and deployment of BPEL processes to JBoss SOA-Platform version 5.1 and earlier.
 OLD_BPEL_FACET_TEMPLATE_LABEL=JBoss BPEL 2.0 Project (for SOA-P v5.1 and earlier)
+
+# This must be updated if the bpel publisher expects to be functional for all jboss server types
+AllJBTServerTypes=org.jboss.ide.eclipse.as.32,org.jboss.ide.eclipse.as.40,org.jboss.ide.eclipse.as.42,org.jboss.ide.eclipse.as.50,org.jboss.ide.eclipse.as.51,org.jboss.ide.eclipse.as.60,org.jboss.ide.eclipse.as.70,org.jboss.ide.eclipse.as.71,org.jboss.ide.eclipse.as.eap.43,org.jboss.ide.eclipse.as.eap.50,org.jboss.ide.eclipse.as.eap.60,org.jboss.ide.eclipse.as.eap.61,org.jboss.ide.eclipse.as.wildfly.80,org.jboss.ide.eclipse.as.systemCopyServer
+
